XML exercises and transcription of letters
There were 2 letters to be transcribed. I was not sure if I needed to create 2 separate txt files or not, so I transcribed both in only one file. I also noticed that the description of one of my letters, the first one, in the BookandByte page, is incorrect. It says: "Letter from E. McKnight Kauffer to Leonard Woolf" when actually it is the other way around: "Letter from the Leonard Woolf to E. McKnight Kauffer".
I had difficulties linking the DTD schema, or the rng file to the xml literatureList one. It says that my document type declaration is still not well-formed. This is something I would like some help with in the next class.